requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
flake8==7.1.2
fonttools==4.56.0
frozenlist==1.5.0
idna==3.10
iniconfig==2.1.0
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.1
mccabe==0.7.0
multidict==6.2.0
mypy-extensions==1.0.0
numpy==2.2.4
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
propcache==0.3.1
pycodestyle==2.12.1
pydantic==2.11.1
pydantic_core==2.33.0
pyflakes==3.2.0
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==5.2.1
requests==2.32.3
rq==2.2.0
rq-dashboard>=0.6.1
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
threadpoolctl==3.6.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
yarl==1.18.3
langchain>=0.1.0
langchain-anthropic>=0.0.1
anthropic>=0.7.0
